% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustTSS.R
\name{adjustTSS}
\alias{adjustTSS}
\title{A function to adjust for near-site clusters for single-end data}
\usage{
adjustTSS(
  anno_path,
  clusters,
  mappedTSS,
  learning_max_dist = 600,
  learning_min_dist = 0,
  learning_methods = "weighted_mean",
  gene_of_interest = NULL,
  gene_extension = 500,
  ncore = 1
)
}
\arguments{
\item{anno_path}{The path to the genome annotation file, which should be in
the .gtf format.}

\item{clusters}{A data.frame containing the near-site clusters regions to be
adjusted. The required columns are seqnames, start, end, strand.}

\item{mappedTSS}{A data.frame containing the mapped TSSs based on the
corresponding single-end data. The required columns are seqnames, tss,
strand, counts.}

\item{learning_max_dist}{An integer to control the adjustment learning
process. Any mapped TSSs which are outside of this range (bp) in the
downstream of an annotated TSS will be ignored in the adjustment learning.
The default is 600(bp).}

\item{learning_min_dist}{An integer to control the adjustment learning
process. Any mapped TSSs which are within this range (bp) in the
downstream of an annotated TSS will be ignored in the adjustment learning.
The default is 0(bp).}

\item{learning_methods}{A character to specify the method used to obtain
the adjustment distance. The input has to be either "weighted_mean" or
"median". The default is set to "weighted_mean".}

\item{gene_of_interest}{A character or a vector of characters, specifying
the genes that are of interest in terms of predicting TSS clusters.}

\item{gene_extension}{An integer specifying the length (bp) of the extended region
before the annotated 5' of a gene when predicting TSS clusters.
Defaults to 500 (bp).}

\item{ncore}{An integer specifying the number of cores used for computation.
Defaults to 1.}
}
\value{
\code{adjustTSS} outputs a list containing two elements.
1) learnt_adj_dist is an integer specifying the learnt adjustment distances;
2) adj_TSS_clusters is a data.frame for the adjusted TSS clusters. It has
these columns: seqnames, tss5prime, width, strand, org_start, org_end.
tss5prime means the location for the 5' end of the adjusted TSS clusters,
and width specifies their lengths. org_start and org_end specify the
corresponding genomic regions before adjustment.
}
\description{
A function to adjust for near-site clusters for single-end data
}
