% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountToUsage.R
\name{CountToUsage}
\alias{CountToUsage}
\title{Convert TSS count to TSS usage}
\usage{
CountToUsage(tss_count_matrix, num_cell = 30)
}
\arguments{
\item{tss_count_matrix}{A TSS count matrix.}

\item{num_cell}{An integer specifying the threshold used to filter out
TSSs that are lowly expressed. A TSS must be expressed in this
many of cells for a single sample. Otherwise it will be removed.
Defaults to 30.}
}
\value{
\code{CountToUsage} outputs a matrix of usage for the
merged TSS by cell. This matrix is stored in data.table format.
}
\description{
Convert TSS count to TSS usage
}
\author{
Shiwei Fu, \email{shiwei.fu@email.ucr.edu}
}
