% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUtest.R
\name{DUtest}
\alias{DUtest}
\title{DU TSS test}
\usage{
DUtest(
  mat,
  col_meta,
  agg_level = "bulk",
  full_model = cbind(TSS_counts, gene_counts - TSS_counts) ~ condition + (1 | sampleID),
  base_model = cbind(TSS_counts, gene_counts - TSS_counts) ~ (1 | sampleID),
  ncore = 1
)
}
\arguments{
\item{mat}{A data.matrix for the count data of TSS clusters.
The rows are TSS clusters and the columns are cells.}

\item{col_meta}{A data.frame containing the column meta information for "mat".
It has to contain at least these columns: barcode, condition, sampleID.}

\item{agg_level}{A string to specify the DU test model ("bulk" or "cell").
"cell" means the DU test are conducted at single-cell level,
"bulk"means the DU test are conducted at the pseudo-bulk level.
The defult is "bulk."}

\item{full_model}{A formula for the full model. The default is
cbind(TSS_counts,gene_counts - TSS_counts) ~ condition + (1|sampleID).}

\item{base_model}{A formula for the base model. The default is
cbind(TSS_counts,gene_counts - TSS_counts) ~ (1|sampleID).}

\item{ncore}{An integer specifying the number of cores used for computation.
Defaults to 1.}
}
\value{
\code{DUtest} returns a data.frame, containing the test results for
each TSS cluster.
}
\description{
DU TSS test
}
