% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMappedTSS.R
\name{makeMappedTSS}
\alias{makeMappedTSS}
\title{Make Mapped TSS table for a sample}
\usage{
makeMappedTSS(
  anno_path,
  bam_path,
  gene_of_interest = NULL,
  if_paired = TRUE,
  gene_extension = 500,
  ncore = 1
)
}
\arguments{
\item{anno_path}{The path to the genome annotation file, which should be in
the .gtf format.}

\item{bam_path}{The path to the .bam, which should be stored under the same
directory as its .bai file.}

\item{gene_of_interest}{A character or a vector of characters, specifying
the gene names that are of interest in terms of predicting Mapped TSS. The
input gene names should be same as those in the genome annotation
(see anno_path). The default is NULL, which means we need to check all genes
provided by the genome annotation (see anno_path).}

\item{if_paired}{FALSE/TRUE, specifying if we have paired-end (on-site) data
(TRUE) or single-end (near-site) data (FALSE). Default set to TRUE.}

\item{gene_extension}{An integer specifying the length of the extended region
before the annotated 5' of a gene when extracting mapped TSS. Defaults to 500 (bp).}

\item{ncore}{An integer specifying the number of cores used for computation.
Defaults to 1.}
}
\value{
\code{makeMappedTSS} outputs a data.frame of Mapped TSS table, which
contains four columns: seqnames, tss (the location of the Mapped TSS),
strand and counts (the count of reads aligned with the Mapped TSS)
}
\description{
Make Mapped TSS table for a sample
}
