% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTSS.R
\name{mergeTSS}
\alias{mergeTSS}
\title{Merge TSS clusters for multiple samples}
\usage{
mergeTSS(
  unmerged_TSS_cluster_list,
  merge_type = "disjoin",
  if_paired = TRUE,
  ncore = 1
)
}
\arguments{
\item{unmerged_TSS_cluster_list}{A list of data.frames. Each data.frame
contains the TSS clusters for one sample. For paired-end data, each
data.frame has these columns: seqnames, start, end and strand. For TSS
clusters predicted from singe-end data, each data.frame has these columns:
seqnames, tss5prime, width, strand, org_start and org_end.}

\item{merge_type}{A character specifying the method used to merge
TSS clusters. There are two methods: "disjoint" and "reduce". Please refer
to \link[GenomicRanges:disjoin]{disjoin},
and \link[GenomicRanges:reduce]{reduce}
for more information. As for single-end data, only "disjoint"
is supported. Defaults to "disjoint".}

\item{if_paired}{FALSE/TRUE, specifying if we have paired-end data
(TRUE) or single-end data (FALSE). Default set to TRUE.}

\item{ncore}{An integer specifying the number of cores used for computation.
This functionality is disabled for the paired-end data. Defaults to 1.}
}
\value{
For paired-end data, \code{mergeTSS} outputs a data.frame, which
stores the unified TSS clusters across samples. The data.frame contains four
columns: seqnames, start, end and strand. For single-end data, the output is
a list, which contains two objects. 1) A data.frame containing the unified TSS
clusters across samples (columns: seqnames, tss5prime,
width and strand); 2) A list of data.frames. Each of them stores the
relations between unified TSS clusters and their corresponding
near-end genomic regions for a sample.
}
\description{
Merge TSS clusters for multiple samples
}
