% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyTSS.R
\name{quantifyTSS}
\alias{quantifyTSS}
\title{Quantify TSS clusters' expression levels for each cell}
\usage{
quantifyTSS(
  anno_path,
  tss_clusters,
  bam_path,
  cell_meta,
  gene_of_interest = NULL,
  reads_percentage = 0.05,
  read_number = 10,
  gene_extension = 500,
  if_paired = TRUE,
  ncore = 1
)
}
\arguments{
\item{anno_path}{The path to the genome annotation file, which should be in
the .gtf format.}

\item{tss_clusters}{A data.frame specifies the TSS clusters to be quantified
for this sample. For paired-end data, these columns should be included:
seqnames, start, end and strand. For single-end data, these column should be
included: seqnames, tss5prime, width, strand, org_start, org_end.}

\item{bam_path}{The path to the .bam, which should be stored under the same
directory as its .bai file.}

\item{cell_meta}{a data.frame specifies the meta information for each cell.
"barcode" and "sampleID" must be included as columns.
"barcode" specifies the barcodes for each of the cells being sequenced
in this sample; "sampleID" is a unique ID for each sample.
Other information are also suggested to be included
in this `data.frame` as well, such as cell types or the
sample's other information (age, gender, etc.).}

\item{gene_of_interest}{A character or a vector of characters, specifying
the gene names that are of interests. The input gene names should be same
as those in the genome annotation (see anno_path).
The default is NULL, which means we need to check all genes provided by the
genome annotation (see anno_path).}

\item{reads_percentage}{A percentage specifying the minim percentage of
reads on a gene should be assigned to a valid TSS clusters within this sample.
This is used to filter out lowly expressed TSS clusters.
Defaults to 0.05.}

\item{read_number}{An integer specifying the minim number of reads that
should be assigned to a valid TSS clusters within this sample.
This is used to filter out lowly expressed TSS clusters.Defaults to 50.}

\item{gene_extension}{An integer specifying the length of the extended region
before the annotated 5' of a gene when quantifying TSS. Defaults to 500 (bp).}

\item{if_paired}{FALSE/TRUE, specifying if we have paired-end data (TRUE)
or single-end data (FALSE). Default set to TRUE.}

\item{ncore}{An integer specifying the number of cores used for computation.
Defaults to 1.}
}
\value{
\code{quantifyTSS} returns a list containing these two elements:
1) TSS_cluster_counts, a data.table
for the expression of each TSS clusters at cell level. 2) Column_meta, a data.frame
containing the meta information for each of the columns in TSS_cluster_counts.
}
\description{
Quantify TSS clusters' expression levels for each cell
}
