% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSduTest.R
\name{TSSduTest}
\alias{TSSduTest}
\title{DU TSS test}
\usage{
TSSduTest(
  DU_input,
  sample_label,
  permutation_times = 2000,
  join_type = "inner",
  ncore = 1
)
}
\arguments{
\item{DU_input}{A list of TSS usage matrices.
The order the matrices should match the sample label.}

\item{sample_label}{A vector of characters specifying the condition
labels for the input matrices. The number of input matrices and vector
of labels should be of the same length.}

\item{permutation_times}{An integer specifying the permutation times
for DU test. Defaults to 2000.}

\item{join_type}{An character, either being "inner" or "outer",
specifying how we unify the merged TSS regions across input usage matrices.
"inner" means we only keep the TSS regions that are shared
by all input matrices for differential usage test.
"outer" means we keep all TSS regions for all input matrices.}

\item{ncore}{An integer specifying the number of cores used. Defaults to 1.}
}
\value{
\code{TSSduTest} outputs the DU test results and the
design matrix used for DU test.
}
\description{
DU TSS test
}
\author{
Shiwei Fu, \email{shiwei.fu@email.ucr.edu}
}
