% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSpredict.R
\name{TSSpredict}
\alias{TSSpredict}
\title{Splicing-aware TSS prediction from .bam file}
\usage{
TSSpredict(
  read_assignmnet_dist = 200,
  gene_extension = 500,
  tss_filtering_distance = 150,
  reads_percentage = 0.05,
  read_number = 50,
  gene_of_interest = NULL,
  if_paired = TRUE,
  ncore = 1,
  bam_path,
  anno_path,
  tss_path,
  barcodes_path
)
}
\arguments{
\item{read_assignmnet_dist}{An integer specifying the threshold distance we
use to assign reads to TSS candidates. A read will be assigned to a
TSS candidate, if the distance from the read to the TSS candidate is
smaller than this threshold. Defaults to 200.}

\item{gene_extension}{An integer specifying the length of the extended region
before the annotated 5' of a gene when predicting TSS. Defaults to 500.}

\item{tss_filtering_distance}{An integer specifying the threshold distance used
to filter out false TSS candidates related to mRNA splicing.
This should be set to the read length of the sequence
technology. Defaults to 150.}

\item{reads_percentage}{A percentage specifying the minim percentage of
reads on a gene should be assigned to a correct TSS candidates.
This is used to filter out false TSS candidate due to background noise.
Defaults to 0.05.}

\item{read_number}{An integer specifying the minim number of reads that
should be assigned to a correct TSS candidate.
This is used to filter out false TSS candidate due to background noise.
Defaults to 50.}

\item{gene_of_interest}{A character or a vector of characters, specifying
the genes that are of interest in terms of predicting TSS.
The default is NULL, which means we need to check all genes provided by the
the genome annotation (see anno_path).}

\item{if_paired}{FALSE/TRUE, specifying if we have paired-end data (TRUE)
or single-end data (FALSE). Default set to TRUE.}

\item{ncore}{An integer specifying the number of cores used for computation.
Defaults to 1.}

\item{bam_path}{The path to the .bam, which should be stored under the same
directory as its .bai file.}

\item{anno_path}{The path to the genome annotation file, which should be in
the .gtf format.}

\item{tss_path}{The path to the .txt file that stores the information
of TSS candidates}

\item{barcodes_path}{The path to the barcode file, which should be a .csv
file with its first column being the barcode and no header.}
}
\value{
\code{TSSprediction} outputs a matrix of read counts for the
predicted TSS by cell. This matrix is stored in data.table format.
}
\description{
Splicing-aware TSS prediction from .bam file
}
\author{
Shiwei Fu, \email{shiwei.fu@email.ucr.edu}
}
