% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMatrixNameSame.R
\name{SetMatrixNameSame}
\alias{SetMatrixNameSame}
\title{Prepare for the input of DU test}
\usage{
SetMatrixNameSame(
  Quantified_TSS_list,
  join_type = "inner",
  exp_level = 0,
  remove_oneTSS_gene = TRUE
)
}
\arguments{
\item{Quantified_TSS_list}{A list of the outputs from the
"quantifyTSS" function. Each element of the list is a single output.}

\item{join_type}{a character, specifies how we prepare the TSS clusters
for the DU test. A TSS cluster is not necessarily
expressed in every sample. If we set the join_type to "inner",
we only consider TSS clusters that are expressed in all of the
samples, which is a conservative choice. If we set the
join_type to “other”, we test TSS clusters that
are expressed in more than a certain percentage of samples.
This percentage is controlled by parameter `exp_level`.
This choice is more sensitive in finding DU TSS. The default is "inner".}

\item{exp_level}{A percentage. When `join_type` is set to "inner,"
this parameter specifies the minimum percentage of cells within
a sample (as indicated by "sampleID" in the column metadata)
required for a TSS cluster to be considered valid.
When `join_type` is other,
it specifies the minimum percentage of samples
in which a TSS cluster must be expressed to be considered valid.
Default is 0, indicating we will not remove TSS clusters based on
expression levels.}

\item{remove_oneTSS_gene}{TRUE/FALSE, specifying if the genes with only one
TSS cluster will be removed. Default is TRUE, indicating the one-TSS
gene will be removed.}
}
\value{
\code{setMatrixNameSame} returns a list containing these two elements:
1) TSS_count_matrix, a data.matrix specifying the TSS cluster counts
ready for DU test; 2) col_meta, a column meta data.frame compatible with
TSS_count_matrix, which is ready for DU test.
}
\description{
Prepare for the input of DU test
}
