% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMatrixNameSame.R
\name{SetMatrixNameSame}
\alias{SetMatrixNameSame}
\title{Prepare for the input of DU test}
\usage{
SetMatrixNameSame(
  Quantified_TSS_list,
  exp_level = 0,
  remove_oneTSS_gene = TRUE,
  join_type = "inner"
)
}
\arguments{
\item{Quantified_TSS_list}{A list of the outputs from the
"quantifyTSS" function. Each element of the list is a single output.}

\item{exp_level}{A percentage specifying, within a sample
(specified by "sampleID" in the column meta data) the minim percentage of
cells that should have a valid TSS cluster expressed. Default is 0,
indicating we will not remove TSS clusters based on expression levels.}

\item{remove_oneTSS_gene}{TRUE/FALSE, specifying if the genes with only one
TSS cluster will be removed. Default is TRUE, indicating the one-TSS
gene will be removed.}

\item{join_type}{a character, specifies how we prepare the TSS clusters
for the DU test. A TSS cluster is not necessarily
expressed in every sample. If we set the join_type to "inner",
we only consider TSS clusters that are expressed in all of the
samples, which is a conservative choice. If we set the
join_type to “outer”, we test TSS regions that
are expressed in as few as only one sample. This choice is more
sensitive in finding DU TSS. The default is "inner".}
}
\value{
\code{setMatrixNameSame} returns a list containing these two elements:
1) TSS_count_matrix, a data.matrix specifying the TSS cluster counts
ready for DU test; 2) col_meta, a column meta data.frame compatible with
TSS_count_matrix, which is ready for DU test.
}
\description{
Prepare for the input of DU test
}
