% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSduTest.R
\name{TSSduTest}
\alias{TSSduTest}
\title{DU TSS test}
\usage{
TSSduTest(
  DU_input,
  sample_label,
  permutation_times = 2000,
  join_type = "inner",
  ncore = 1
)
}
\arguments{
\item{DU_input}{A list of TSS usage matrices output from \code{CountToUsage}.
The order the matrices should match the sample labels in "sample_label".}

\item{sample_label}{A vector of characters specifying the condition
labels for the input matrices. The number of input matrices and vector
of labels should be of the same length.}

\item{permutation_times}{An integer specifying the permutation times
for DU test. Defaults to 2000.}

\item{join_type}{a character, specifies how we unify the TSS regions
in the input TSS usage matrices. A TSS region is not necessarily
expressed in every sample. If we set the join_type to "inner",
 we only consider TSS regions that are expressed in all of the
 samples, which is a conservative choice. If we set the
 join_type to “outer”, we test TSS regions that
 are expressed in as few as only one sample. This choice is more
 sensitive in finding DU TSS. The default is "inner".}

\item{ncore}{An integer specifying the number of cores used for computation.
Defaults to 1.}
}
\value{
\code{TSSduTest} outputs the DU results and the
design matrix used for DU.
}
\description{
DU TSS test
}
\author{
Shiwei Fu, \email{shiwei.fu@email.ucr.edu}
}
