% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSmerge.R
\name{TSSmerge}
\alias{TSSmerge}
\title{TSS merging for TSS count matrices of multiple samples}
\usage{
TSSmerge(tss_matrices, tss_filtering_distance, ncore = 1)
}
\arguments{
\item{tss_matrices}{A list of TSS count matrices output
from \code{TSSpredict}. Each element in the list should a matrix.}

\item{tss_filtering_distance}{An integer specifying the threshold distance
to merge TSS candidates into a same TSS region.
This should be set to the same as in \code{TSSpredict}. Defaults to 1.}

\item{ncore}{An integer specifying the number of cores used for computation.
Defaults to 1.}
}
\value{
\code{TSSmerge} outputs a matrix of read counts for the
merged TSS regions by cell. This matrix is stored in data.table format.
}
\description{
TSS merging for TSS count matrices of multiple samples
}
\author{
Shiwei Fu, \email{shiwei.fu@email.ucr.edu}
}
