% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSpredict.R
\name{TSSpredict}
\alias{TSSpredict}
\title{Splicing-aware TSS prediction from .bam file}
\usage{
TSSpredict(
  read_assignmnet_dist = 200,
  gene_extension = 500,
  tss_filtering_distance = 150,
  reads_percentage = 0.05,
  read_number = 50,
  gene_of_interest = NULL,
  if_paired = TRUE,
  oneTSSgene_path = NULL,
  ncore = 1,
  bam_path,
  anno_path,
  tss_path,
  barcodes_path
)
}
\arguments{
\item{read_assignmnet_dist}{An integer specifying the threshold distance we
use to assign reads to TSS candidates. Defaults to 200.}

\item{gene_extension}{An integer specifying the extension before the
annotated 5' of a gene. Defaults to 500.}

\item{tss_filtering_distance}{An integer specifying the threshold distance
for TSS candidates that should be considered from the same genuine TSS.
This should be set to the read length of the sequence technology.
Defults to 150.}

\item{reads_percentage}{A percentage specifying the minim percentage of
reads on a gene should be assigned to a qualified TSS candidates.
Defaults to 0.05.}

\item{read_number}{An integer specifying the minim number of reads that
should be assigned to a qualified TSS candidate. Defaults to 50.}

\item{gene_of_interest}{A character or a vector of characters, specifying
the genes that are of interest in terms of predicting TSS.
The default is NULL, which means we need to check all genes provided by the
the genome annotation (see anno_path). Otherwise provide a subset of gene IDs
that is in the provided annotation.}

\item{if_paired}{FALSE/TRUE, specifying if we have paired-end data (TRUE)
or single-end data (FALSE). Default set to TRUE.}

\item{oneTSSgene_path}{The path to a list of gene that are used in the
learning step for single-end data. This only need to be specified if the
data is single-end. Defaults to NULL.}

\item{ncore}{An integer specifying the number of cores used. Defaults to 1.}

\item{bam_path}{The path to the .bam, which should be stored under the same
directory as its .bai file.}

\item{anno_path}{The path to the genome annotation file, which should be in
the .gtf format.}

\item{tss_path}{The path to the .txt file that stores the information
of TSS candidates}

\item{barcodes_path}{The path to the barcode file, which should be a .csv
file with its first column being the barcode and no header.}
}
\value{
\code{TSSprediction} outputs a matrix of read counts for the
predicted TSS by cell. This matrix is stored in data.table format.
}
\description{
Splicing-aware TSS prediction from .bam file
}
\author{
Shiwei Fu, \email{shiwei.fu@email.ucr.edu}
}
